+++ app/lib/classic_game/engine.rb
@@ -64,21 +64,25 @@ module ClassicGame
           errors << error_text
         end
 
+        valid_movement_types = %w[patrol triggered]
         %w[npcs creatures].each do |collection|
+          label = collection.singularize.capitalize
           (world_data[collection] || {}).each do |entity_id, entity_def|
             next unless entity_def.is_a?(Hash) && entity_def["movement"]
 
             movement = entity_def["movement"]
-            unless %w[patrol triggered].include?(movement["type"])
-              errors << "#{collection.singularize.capitalize} '#{entity_id}' has unknown movement type '#{movement['type']}'."
+            unless valid_movement_types.include?(movement["type"])
+              errors << "#{label} '#{entity_id}' has unknown movement type " \
+                        "'#{movement['type']}'."
             end
 
             if movement["type"] == "patrol" && (movement["route"].blank? || !movement["route"].is_a?(Array))
-              errors << "#{collection.singularize.capitalize} '#{entity_id}' patrol movement requires a non-empty route array."
+              errors << "#{label} '#{entity_id}' patrol movement " \
+                        "requires a non-empty route array."
             end
 
             if movement["type"] == "triggered" && movement["destination"].blank?
-              errors << "#{collection.singularize.capitalize} '#{entity_id}' triggered movement requires a destination."
+              errors << "#{label} '#{entity_id}' triggered movement requires a destination."
             end
           end
         end
