+++ app/lib/classic_game/movement_processor.rb
@@ -11,16 +11,16 @@ module ClassicGame
         (world["npcs"] || {}).each do |npc_id, npc_def|
           next unless npc_def.is_a?(Hash) && npc_def["movement"]
 
-          msg = process_entity(game: game, entity_type: "npc", entity_id: npc_id,
-                               entity_def: npc_def, player_room: player_room)
+          ctx = build_context(game, "npc", npc_id, npc_def, player_room)
+          msg = process_entity(ctx)
           messages << msg if msg
         end
 
         (world["creatures"] || {}).each do |creature_id, creature_def|
           next unless creature_def.is_a?(Hash) && creature_def["movement"]
 
-          msg = process_entity(game: game, entity_type: "creature", entity_id: creature_id,
-                               entity_def: creature_def, player_room: player_room)
+          ctx = build_context(game, "creature", creature_id, creature_def, player_room)
+          msg = process_entity(ctx)
           messages << msg if msg
         end
 
@@ -29,23 +29,31 @@ module ClassicGame
 
       private
 
-        def process_entity(game:, entity_type:, entity_id:, entity_def:, player_room:)
-          movement = entity_def["movement"]
+        def build_context(game, entity_type, entity_id, entity_def, player_room)
+          {
+            game: game, entity_type: entity_type, entity_id: entity_id,
+            entity_def: entity_def, movement: entity_def["movement"],
+            player_room: player_room
+          }
+        end
 
-          case movement["type"]
+        def process_entity(ctx)
+          case ctx[:movement]["type"]
           when "patrol"
-            process_patrol(game: game, entity_type: entity_type, entity_id: entity_id,
-                           entity_def: entity_def, movement: movement, player_room: player_room)
+            process_patrol(ctx)
           when "triggered"
-            process_triggered(game: game, entity_type: entity_type, entity_id: entity_id,
-                              entity_def: entity_def, movement: movement, player_room: player_room)
+            process_triggered(ctx)
           end
         end
 
-        def process_patrol(game:, entity_type:, entity_id:, entity_def:, movement:, player_room:)
+        def process_patrol(ctx)
+          game = ctx[:game]
+          entity_type = ctx[:entity_type]
+          entity_id = ctx[:entity_id]
+          movement = ctx[:movement]
+
           # Skip if entity no longer exists in any room (e.g. defeated creature)
-          current_entity_room = find_entity_room(game, entity_type, entity_id)
-          return nil unless current_entity_room
+          return nil unless find_entity_room(game, entity_type, entity_id)
 
           state = game.movement_state(entity_type, entity_id)
           route = movement["route"]
@@ -61,19 +69,16 @@ module ClassicGame
 
           if counter >= stay_duration
             unless_rooms = movement["unless_player_in"] || []
-            if unless_rooms.include?(player_room) && current_room == player_room
+            if unless_rooms.include?(ctx[:player_room]) && current_room == ctx[:player_room]
               game.update_movement_state(entity_type, entity_id,
                                          { "step" => step, "move_counter" => counter })
               return nil
             end
 
             next_step = (step + 1) % route.length
-            next_stop = route[next_step]
-            destination = next_stop["room"]
+            destination = route[next_step]["room"]
 
-            message = move_entity(game: game, entity_type: entity_type, entity_id: entity_id,
-                                  entity_def: entity_def, from_room: current_room,
-                                  to_room: destination, player_room: player_room, movement: movement)
+            message = move_entity(ctx, from_room: current_room, to_room: destination)
 
             game.update_movement_state(entity_type, entity_id,
                                        { "step" => next_step, "move_counter" => 0 })
@@ -85,7 +90,12 @@ module ClassicGame
           nil
         end
 
-        def process_triggered(game:, entity_type:, entity_id:, entity_def:, movement:, player_room:)
+        def process_triggered(ctx)
+          game = ctx[:game]
+          entity_type = ctx[:entity_type]
+          entity_id = ctx[:entity_id]
+          movement = ctx[:movement]
+
           state = game.movement_state(entity_type, entity_id)
           return nil if state["triggered"]
 
@@ -97,9 +107,7 @@ module ClassicGame
           return nil unless current_room
           return nil if current_room == destination
 
-          message = move_entity(game: game, entity_type: entity_type, entity_id: entity_id,
-                                entity_def: entity_def, from_room: current_room,
-                                to_room: destination, player_room: player_room, movement: movement)
+          message = move_entity(ctx, from_room: current_room, to_room: destination)
 
           game.update_movement_state(entity_type, entity_id, { "triggered" => true })
           message
@@ -117,25 +125,27 @@ module ClassicGame
           nil
         end
 
-        def move_entity(game:, entity_type:, entity_id:, entity_def:, from_room:, to_room:, player_room:, movement:)
+        def move_entity(ctx, from_room:, to_room:)
+          game = ctx[:game]
+          entity_type = ctx[:entity_type]
           collection_key = entity_type == "npc" ? "npcs" : "creatures"
-          entity_name = entity_def["name"] || entity_id
+          entity_name = ctx[:entity_def]["name"] || ctx[:entity_id]
 
           old_room_state = game.room_state(from_room).dup
-          old_room_state[collection_key] = (old_room_state[collection_key] || []) - [entity_id]
+          old_room_state[collection_key] = (old_room_state[collection_key] || []) - [ctx[:entity_id]]
           old_room_state["modified"] = true
           game.update_room_state(from_room, old_room_state)
 
           new_room_state = game.room_state(to_room).dup
           new_room_state[collection_key] ||= []
-          new_room_state[collection_key] << entity_id
+          new_room_state[collection_key] << ctx[:entity_id]
           new_room_state["modified"] = true
           game.update_room_state(to_room, new_room_state)
 
-          if from_room == player_room
-            movement["depart_text"] || "#{entity_name} leaves."
-          elsif to_room == player_room
-            movement["arrive_text"] || "#{entity_name} arrives."
+          if from_room == ctx[:player_room]
+            ctx[:movement]["depart_text"] || "#{entity_name} leaves."
+          elsif to_room == ctx[:player_room]
+            ctx[:movement]["arrive_text"] || "#{entity_name} arrives."
           end
         end
     end
